chore: refactor ListTable
